Moved old routers to directory to avoid confusion
